Add doc and rename IDecoder to AbstractDecoder

+++ src/model/Coordinate.h
@@ -1,3 +1,9 @@
+/**
+ * @file Coordinate.h
+ * @author Baudouin Feildel <baudouin.feildel@st.com>
+ * @copyright 2016, STMicroelectronics, All rights reserved.
+ */
+
 #ifndef TESEO_HAL_MODEL_COORDINATE
 #define TESEO_HAL_MODEL_COORDINATE
 
@@ -9,7 +15,10 @@ class DegreeMinuteCoordinate;
 
 class DecimalDegreeCoordinate;
 
-enum class CoordinateDirection : uint
+/**
+ * @brief      Coordinate direction enumeration
+ */
+enum class CoordinateDirection : uint8_t
 {
 	UNKNOWN,
 	NORTH,
@@ -18,17 +27,43 @@ enum class CoordinateDirection : uint
 	WEST,
 };
 
+/**
+ * @brief      Parse ascii char to coordinate direction
+ *
+ * @param[in]  byte  The byte to parse
+ *
+ * @return     The coordinate direction, or UNKNOWN if character is not N, S, E nor W.
+ */
 CoordinateDirection CoordinateDirectionParse(uint8_t byte);
 
+/**
+ * @brief      Coordinate interface
+ */
 class ICoordinate
 {
 public:
+	/**
+	 * @brief      Convert coordinate to degree minute coordinate
+	 *
+	 * @return     New degree minute coordinate object
+	 */
 	virtual DegreeMinuteCoordinate asDegreeMinute() const = 0;
+
+	/**
+	 * @brief      Convert coordinate to decimal degree coordinate
+	 *
+	 * @return     New decimal degree coordinate
+	 */
 	virtual DecimalDegreeCoordinate asDecimalDegree() const = 0;
 
 	virtual ~ICoordinate() { }
 };
 
+/**
+ * @brief      Degree minute coordinate class
+ * @details    A degre minute coordinate store the degree as an integer, minute as decimal minutes
+ * (double) and the coordinate direction.
+ */
 class DegreeMinuteCoordinate:
 	public ICoordinate
 {
@@ -46,6 +81,12 @@ public:
 
 	DegreeMinuteCoordinate(int degree, double minute, CoordinateDirection direction);
 
+	/**
+	 * @brief      Create coordinate from ascii string
+	 *
+	 * @param[in]  coordinate  The coordinate string
+	 * @param[in]  direction   The direction character
+	 */
 	DegreeMinuteCoordinate(const ByteVector & coordinate, uint8_t direction);
 
 	int getDegree() const { return degree; }
@@ -59,6 +100,11 @@ public:
 	DecimalDegreeCoordinate asDecimalDegree() const;
 };
 
+/**
+ * @brief      Decimal degree coordinate class
+ * @details    A decimal degree coordinate store the coordinate as a double. Direction is indicated
+ * by the sign.
+ */
 class DecimalDegreeCoordinate:
 	public ICoordinate
 {
@@ -74,6 +120,11 @@ public:
 
 	DecimalDegreeCoordinate(double coordinate);
 
+	/**
+	 * @brief      Create coordinate from ascii string
+	 *
+	 * @param[in]  coordinate  The coordinate string
+	 */
 	DecimalDegreeCoordinate(const ByteVector & coordinate);
 
 	double value() const { return coordinate; }
